Rename LoaDomain to ACCRLookupService.
